docs: add docblock comments to all libdave class variables (#1286)

+++ src/dpp/dave/frame_processors.h
@@ -73,7 +73,7 @@ uint8_t serialize_unencrypted_ranges(const ranges& unencrypted_ranges, uint8_t* 
  * @param unencrypted_ranges unencrypted ranges to write to
  * @return size of unencrypted ranges written
  */
-uint8_t deserialize_unencrypted_ranges(const uint8_t*& read_at, const size_t buffer_size, ranges& unencrypted_ranges);
+uint8_t deserialize_unencrypted_ranges(const uint8_t*& read_at, const uint8_t buffer_size, ranges& unencrypted_ranges);
 
 /**
  * @brief Validate unencrypted ranges
@@ -183,13 +183,44 @@ private:
 	 */
 	void add_ciphertext_bytes(const uint8_t* data, size_t size);
 
+	/**
+	 * @brief True if frames are encrypted
+	 */
 	bool encrypted{false};
+
+	/**
+	 * @brief Original size
+	 */
 	size_t original_size{0};
+
+	/**
+	 * @brief AEAD tag
+	 */
 	array_view<const uint8_t> tag;
+
+	/**
+	 * @brief Truncated nonce
+	 */
 	truncated_sync_nonce truncated_nonce;
+
+	/**
+	 * @brief Unencrypted parts of the frames
+	 */
 	ranges unencrypted_ranges;
+
+	/**
+	 * @brief additional authenticated data
+	 */
 	std::vector<uint8_t> authenticated;
+
+	/**
+	 * @brief Ciphertext
+	 */
 	std::vector<uint8_t> ciphertext;
+
+	/**
+	 * @brief Plaintext
+	 */
 	std::vector<uint8_t> plaintext;
 
 	/**
@@ -283,11 +314,34 @@ public:
 	void add_encrypted_bytes(const uint8_t* bytes, size_t size);
 
 private:
+	/**
+	 * @brief Codec used to decrypt
+	 */
 	codec frame_codec{codec::cd_unknown};
+
+	/**
+	 * @brief Frame index
+	 */
 	size_t frame_index{0};
+
+	/**
+	 * @brief Unencrypted bytes
+	 */
 	std::vector<uint8_t> unencrypted_bytes;
+
+	/**
+	 * @brief Encrypted bytes
+	 */
 	std::vector<uint8_t> encrypted_bytes;
+
+	/**
+	 * @brief Ciphertext bytes
+	 */
 	std::vector<uint8_t> ciphertext_bytes;
+
+	/**
+	 * @brief Unencrypted ranges that need to be kept plaintext to allow for RTP routing
+	 */
 	ranges unencrypted_ranges;
 
 	/**
